Update AvatarEvaluator.cs
Commented out a number of default shaders that likely are not SPS-I compatible, or generally shouldn't be in use on an avatar since the components that would use them are not whitelisted. Put an issue on the issues page or put in a Pull Request incase you actually needed any of these or any of these actually work. Also Write Defaults being all true or all false doesn't matter in CVR, but VRC avatars expect that behaviour and mixed cases may make things act funky, a warning is in place for this now.